+++ src/Fauna/hft.h
@@ -35,8 +35,8 @@ typedef std::vector<std::shared_ptr<const Hft> > HftList;
  * \f[
  * x = c * M^e
  * \f]
- * - $c$ = \ref coefficient
- * - $e$ = \ref exponent
+ * - \f$c\f$ = \ref coefficient
+ * - \f$e\f$ = \ref exponent
  */
 struct AllometryParameters {
   /// Constructor.
@@ -51,6 +51,28 @@ struct AllometryParameters {
 
   /// Calculate the result of the formula.
   double calc(const double M) const { return coefficient * pow(M, exponent); }
+};
+
+/// Parameters for an allometric relationship with exponent and one point.
+/**
+ * The allometric relationship is \f$ f(M) = c * M^e \f$, where \f$c\f$ is the
+ * coefficient and \f$e\f$ is the \ref exponent. The coefficient is calculated
+ * from a given point \f$(x|y)\f$ with
+ * \f$x = f(y) = c * y^e \iff c = x * y^{-e} \f$.
+ *
+ * Generally, \f$y\f$ is the body mass of an adult male
+ * (\ref Hft::body_mass_male). This is because males are typically larger than
+ * females. Allometric extrapolating from smaller females to larger males would
+ * be more uncertain than the other way round.
+ *
+ * \see \ref calc_allometry()
+ */
+struct GivenPointAllometry {
+  /// Exponent \f$e\f$ in \f$f(M) = c * M^e\f$.
+  double exponent;
+
+  /// Value \f$f(m)\f$ if \f$m\f$ is \ref Hft::body_mass_male.
+  double value_male_adult;
 };
 
 /// Selector for a function of how to calculate whole-body conductance.
@@ -364,7 +386,7 @@ struct Hft {
 
   /** @{ \name "digestion": Digestion-related parameters. */
   /// Parameters for \ref DigestiveLimit::Allometric
-  AllometryParameters digestion_allometric = {0.05, 0.76};
+  GivenPointAllometry digestion_allometric = {0.05, 0.76};
 
   /// Factor to change ruminant digestibility for other digestion types.
   /**